Change double (fp64) to float (fp32) for GPU
to evaluate if using fp32 is faster (obviously less precise, but I'm only interested in O(1%) precision for now)

+++ pySecDec/code_writer/templates/sum_package/name.hpp
@@ -39,7 +39,7 @@ namespace %(name)s
     
     // basic data types
     // --{
-    typedef double real_t;
+    typedef float real_t;
     #ifdef SECDEC_WITH_CUDA
         typedef thrust::complex<real_t> complex_t;
     #else
